fix: rename useFetchConstants to useConstants where it was not changed
