+++ src/hooks/useAddressBook/types.ts
@@ -10,7 +10,7 @@ export type Address = {
 
 export interface UseAddressBookProps {
   // TODO: Add type
-  useFetchConstants: any
+  useConstants: any
   useAccounts: () => UseAccountsReturnType
   useStorage: UseStorageType
   useToasts: () => UseToastsReturnType
